requirements.txt
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
pandas=1.1.5=pypi_0
python=3.8.5=h7579374_1

scikit-learn=0.24.0=pypi_0
scipy=1.5.2=py38h0b6359f_0
torch-cluster=1.5.8=pypi_0
torch-geometric=1.6.3=pypi_0
torch-scatter=2.0.5=pypi_0
torch-sparse=0.6.8=pypi_0
torch-spline-conv=1.2.0=pypi_0

